Separate subprocess module
